LPS-87529 regen
