+++ modules/apps/headless/headless-document-library/headless-document-library-api/src/main/java/com/liferay/headless/document/library/resource/v1_0/DocumentResource.java
@@ -15,7 +15,6 @@
 package com.liferay.headless.document.library.resource.v1_0;
 
 import com.liferay.headless.document.library.dto.v1_0.Document;
-import com.liferay.headless.document.library.dto.v1_0.Rating;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.kernel.search.Sort;
 import com.liferay.portal.kernel.search.filter.Filter;
@@ -53,12 +52,6 @@ public interface DocumentResource {
 		throws Exception;
 
 	public Document putDocument(Long documentId, MultipartBody multipartBody)
-		throws Exception;
-
-	public Page<Rating> getDocumentsRatingsPage(Long documentId)
-		throws Exception;
-
-	public Rating postDocumentRating(Long documentId, Rating rating)
 		throws Exception;
 
 	public Page<Document> getFolderDocumentsPage(
