+++ modules/apps/headless/headless-document-library/headless-document-library-impl/src/main/java/com/liferay/headless/document/library/internal/resource/v1_0/BaseRatingResourceImpl.java
@@ -19,12 +19,14 @@ import com.liferay.headless.document.library.resource.v1_0.RatingResource;
 import com.liferay.petra.function.UnsafeFunction;
 import com.liferay.portal.kernel.model.Company;
 import com.liferay.portal.vulcan.accept.language.AcceptLanguage;
+import com.liferay.portal.vulcan.pagination.Page;
 import com.liferay.portal.vulcan.util.TransformUtil;
 
 import io.swagger.v3.oas.annotations.tags.Tag;
 import io.swagger.v3.oas.annotations.tags.Tags;
 
 import java.util.Collection;
+import java.util.Collections;
 import java.util.List;
 
 import javax.annotation.Generated;
@@ -34,6 +36,7 @@ import javax.validation.constraints.NotNull;
 import javax.ws.rs.Consumes;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
+import javax.ws.rs.POST;
 import javax.ws.rs.PUT;
 import javax.ws.rs.Path;
 import javax.ws.rs.PathParam;
@@ -48,6 +51,31 @@ import javax.ws.rs.core.UriInfo;
 @Generated("")
 @Path("/v1.0")
 public abstract class BaseRatingResourceImpl implements RatingResource {
+
+	@Override
+	@GET
+	@Path("/documents/{document-id}/ratings")
+	@Produces("application/json")
+	@Tags(value = {@Tag(name = "Rating")})
+	public Page<Rating> getDocumentRatingsPage(
+			@NotNull @PathParam("document-id") Long documentId)
+		throws Exception {
+
+		return Page.of(Collections.emptyList());
+	}
+
+	@Override
+	@Consumes("application/json")
+	@POST
+	@Path("/documents/{document-id}/ratings")
+	@Produces("application/json")
+	@Tags(value = {@Tag(name = "Rating")})
+	public Rating postDocumentRating(
+			@NotNull @PathParam("document-id") Long documentId, Rating rating)
+		throws Exception {
+
+		return new Rating();
+	}
 
 	@Override
 	@DELETE
